Update to AIO Boot v0.9.4.7
- In previous versions, some user encountered an error when using the
"**FAT32 & NTFS**" option on Windows 10 v1703. And one of them has
confirmed that the error is no longer in this version. **Note**: Windows
10 v1703 can mount multiple drives for all partitions on USB.
- Added support for **Norman Rescue Disk** as desired by a user. Maybe
Norman Rescue Disk will not work on NTFS, try FAT32 instead. If you do
not want to use this package, you do not need to update this version.